trigger-period, correct width, pony ORM
This adds a second page with plots similar to the trigger-lifetime
charts which can be found along with the trending files.

Among the changes in the js files is also a function which fetches the
width of a chart from the bootstrap column widths its supposed to
have.

Lastly, this commit chages the parsing of the input root files. It now
uses Pony ORM in order to create a in-memory SQL database. This makes
it easy to csv create files of arbitrary content and pre-filtered at
ease. Performace could be optimized, though...